TINKERPOP-2959 Allowed parsing of parameters with the grammar
This adds functionality similar to GremlinGroovyScriptEngine where bindings can be applied to a script as in g.V(x) where x is a variable that can be bound to a value external to the script. This functionality doesn't offer any particular performance enhancement for the grammar the way that it does for groovy but it does better allow one engine to be swapped for the other which is crucial for a migration away from the groovy engine. This change also opens up better opportunities for translator and gherkin improvements as well.